Special-case Azure DevOps' URLs
Azure DevOps' URL scheme is https://dev.azure.com/<org>/<repo>, and it
is quite common for people to need *different* credentials depending on
the org.

So let's teach Firefox' Containers about this special case, and use
not only the host name but also the org to determine the site key (that
can be associated with a container).

This starts to address
https://github.com/mozilla/multi-account-containers/issues/1254

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>